turn redis.test.js into a generic set of tests that can be run on any backend, and implement TokenStore.ping
